+++ src/main/java/com/jp/testesMockito/config/LocalConfig.java
@@ -4,14 +4,13 @@ import com.jp.testesMockito.domain.User;
 import com.jp.testesMockito.repositories.UserRepository;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
 import java.util.List;
 
 @Configuration
-@Profile("local")
+@Profile("test")
 @RequiredArgsConstructor
 public class LocalConfig {
