+++ src/main/java/com/jp/testesMockito/mapper/UserMapper.java
@@ -3,14 +3,11 @@ package com.jp.testesMockito.mapper;
 import com.jp.testesMockito.domain.User;
 import com.jp.testesMockito.domain.dto.UserDTO;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 
 import java.util.List;
 
 @Mapper(componentModel = "spring")
 public interface UserMapper {
-
-
 
     UserDTO toDTO(User user);
